+++ src/crypto/index.ts
@@ -4,10 +4,8 @@ import subtle from "./subtle";
 
 import * as key from './key'
 
-import { web_key_to_crypto_key } from "./web_key_to_crypto_key";
-
 import * as web from './web'
 
-export { signer, verifier, key, subtle, web_key_to_crypto_key, web }
+export { signer, verifier, key, subtle, web }
